+++ ReleaseNotes.md
@@ -1,6 +1,7 @@
 # 2020.2.xx.xxx-pre-release
 * Even if a union type is [Obsolete] don't bother telling us its cases and case constructors depend on it.
 * In F# code (assumes FSharp.Core referenced) then allow single lower-case letter generic types
+* Exempt methods of F# generated types with `@` in the name from ` ParameterNamesShouldMatchOverriddenMethodRule`
 
 For previous releases, go here -- https://github.com/SteveGilham/Gendarme/blob/master/ReleaseNotes%20-%20Previously.md
 Since this is a release from a fork, [issues should be reported at my related repo](https://github.com/SteveGilham/altcode.fake/issues) that contains a Fake driver for the Gendarme tool, but noted as being against the forked Gendarme tool itself.